python/bridge_app/requirements.txt
flask
flask-cors
pillow
numpy
scipy
tomli; python_version < "3.11"

